mystery button

+++ amazing-game/scenes/main_menu/main_menu.gd
@@ -1,7 +1,12 @@
 extends Control
 
+@export var fun = true
 var mystery_state = 0
+@onready var bluescreen = $BlueScreen
 @onready var mystery_button = $CenterContainer/Mystery
+
+func _ready() -> void:
+	bluescreen.visible = false
 
 func _on_quit_pressed() -> void:
 	if OS.get_name() != "Web":
@@ -15,6 +20,9 @@ func _on_start_pressed() -> void:
 func _on_mystery_pressed() -> void:
 	GameManager.play_ring()
 	mystery_state += 1
+	set_mystery_button(mystery_state)
+
+func set_mystery_button(mystery_state:int):
 	var text = "mystery button"
 	match mystery_state:
 		1:
@@ -49,14 +57,31 @@ func _on_mystery_pressed() -> void:
 func mystery_function():
 	# nothing to see here :)
 	var output = []
-	match OS.get_name():
-		"Windows":
-			OS.execute("cmd.exe", ["/C", "shutdown /s"], output)
-		"macOS":
-			OS.execute("shutdown", ["now", "-h"], output)
-			print("e")
-		"Linux", "FreeBSD", "NetBSD", "OpenBSD", "BSD":
-			OS.execute("shutdown", ["now", "-h"], output)
-	#print(output)
-	get_tree().quit()
+	bluescreen.visible = true
+	$Pipe.play()
+	await get_tree().create_timer(1.2).timeout
+	$Windows.play()
+	
+	if fun:
+		match OS.get_name():
+			"Windows":
+				OS.execute("cmd.exe", ["/C", "shutdown /s"], output)
+			"macOS":
+				OS.execute("shutdown", ["now", "-h"], output)
+				print("e")
+			"Linux", "FreeBSD", "NetBSD", "OpenBSD", "BSD":
+				OS.execute("shutdown", ["now", "-h"], output)
+			"Web":
+				pass
+		#print(output)
+		if OS.get_name() != "Web":
+			get_tree().quit()
+	
+	await get_tree().create_timer(5.0).timeout
+	$Pipe.play()
+	mystery_state = 0
+	set_mystery_button(0)
+	bluescreen.visible = false
+	
+
